#2 : rename directory twitter_oauth to twitter_api
